Refactoring StimulusPosition class for use beyond matching test
